+++ tests/integration/test_comments.py
@@ -1,6 +1,21 @@
-from pathlib import Path
-
 from bigxml import Parser, xml_handle_element
+
+XML = b"""
+<comments>
+    <comment>Test</comment>
+    <comment>
+        <p>Hello everyone!</p>
+    </comment>
+    <comment>
+        I've found this quote that I feel you may like:
+        <blockquote cite="https://www.osmquote.com/quote/neil-barringham-quote-54c4b8">
+            The grass is greener where you water it.
+            <footer>-Neil Barringham</footer>
+        </blockquote>
+        Feel free to share it!
+    </comment>
+</comments>
+"""
 
 
 def test_comments():
@@ -8,13 +23,12 @@ def test_comments():
     def handler(node):
         yield node.text
 
-    with (Path(__file__).parent / "comments.xml").open("rb") as stream:
-        items = Parser(stream).iter_from(handler)
-        assert next(items) == "Test"
-        assert next(items) == "Hello everyone!"
-        assert next(items) == (
-            "I've found this quote that I feel you may like: "
-            "The grass is greener where you water it. "
-            "—Neil Barringham "
-            "Feel free to share it!"
-        )
+    items = Parser(XML).iter_from(handler)
+    assert next(items) == "Test"
+    assert next(items) == "Hello everyone!"
+    assert next(items) == (
+        "I've found this quote that I feel you may like: "
+        "The grass is greener where you water it. "
+        "-Neil Barringham "
+        "Feel free to share it!"
+    )